+++ ensureNoTrailingSlash.go
@@ -1,11 +1,10 @@
 package fileIO
 
+import "strings"
+
 // Returns the string without a trailing slash. If there is none, it is returned as is
 func ensureNoTrailingSlash(text string) string {
 
-	if text[len(text)-1:] == "/" {
-		return text[:len(text)-1]
-	} else {
-		return text
-	}
+	return strings.TrimSuffix(text, "/")
+
 }