+++ CHANGELOG.md
@@ -1,4 +1,10 @@
 # CHANGELOG
 
+## v1.0.1 on 2025-09-19
+
+- Ensure empty strings are returned unchanged on `ensureTrailingSlash`
+- Make use of `strings.HasSuffix` instead of `text[len(text)-1:]`
+
 ## v1.0.0 on 2023-02-14
+
 - initial release